Fixed another type mismatch
git-svn-id: http://svn.drobilla.net/lad/libslv2@104 a436a847-0d15-0410-975c-d299462d15a1

+++ slv2/pluginlist.h
@@ -146,7 +146,7 @@ slv2_list_get_plugin_by_uri(const SLV2List list,
  */
 const SLV2Plugin*
 slv2_list_get_plugin_by_index(const SLV2List list,
-                              unsigned long  index);
+                              size_t         index);
 
 
 /** @} */
